--- packages/addon/CONTRIBUTING.md
@@ -1,26 +1,0 @@
-# How To Contribute
-
-## Installation
-
-* `git clone <repository-url>`
-* `cd @embroider/addon`
-* `yarn install`
-
-## Linting
-
-* `yarn lint:hbs`
-* `yarn lint:js`
-* `yarn lint:js --fix`
-
-## Running tests
-
-* `ember test` – Runs the test suite on the current Ember version
-* `ember test --server` – Runs the test suite in "watch mode"
-* `ember try:each` – Runs the test suite against multiple Ember versions
-
-## Running the dummy application
-
-* `ember serve`
-* Visit the dummy application at [http://localhost:4200](http://localhost:4200).
-
-For more information on using ember-cli, visit [https://ember-cli.com/](https://ember-cli.com/).
