--- packages/addon/app/helpers/ensure-safe-component.js
@@ -1,1 +1,0 @@
-export { EnsureSafeComponentHelper as default } from '@embroider/addon';
