+++ packages/util/addon/index.js
@@ -30,7 +30,7 @@ function handleString(name, thingWithOwner) {
       url:
         'https://github.com/embroider-build/embroider/blob/master/ADDON-AUTHOR-GUIDE.md#when-youre-passing-a-component-to-someone-else',
       until: 'embroider',
-      for: '@embroider/addon',
+      for: '@embroider/util',
       since: '0.27.0',
     }
   );
