+++ packages/util/app/helpers/ensure-safe-component.js
@@ -1,0 +1,1 @@
+export { EnsureSafeComponentHelper as default } from '@embroider/util';
