+++ packages/util/tests/dummy/app/components/colocated-example.hbs
